Make sap.h and sap.c into single header
