+++ tests/CMakeLists.txt
@@ -1,4 +1,6 @@
 add_executable(ctests src/ctests.c)
 add_executable(cpptests src/cpptests.cpp)
-target_link_libraries(ctests sap)
-target_link_libraries(cpptests sap)
+
+# include header files
+target_include_directories(ctests PRIVATE ../include)
+target_include_directories(cpptests PRIVATE ../include)
